+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# Unreleased
+
+* Add RSpec test helpers
+
 # 2.1.1
 
 * Support for a new regex in GOV.UK content schemas